Generate rcxnub and rcxnub_odd from the assembly file
Move one instruction to make it bit exact with the previously included
SREC file, add conditional to also generate the odd parity version.

The unused rcxnub_linuxusb.h and the commented out version in
rcxnub_odd.h were the same. They explicitly set odd parity while the non
commented version just skip changing the register.

On build, if H8300 toolchain is not available, use a pre-built version.
If toolchain is available, the pre-built version is checked to make sure
it is up to date.